Revert "Refactor/collection error fix 수정중"
