+++ src/components/myPage/Collections/Collection.jsx
@@ -21,12 +21,13 @@ const Collection = ({ setImgSrc }) => {
 
   const collectionQueries = useCollectionQueries(collection, { enabled: !!collection });
 
-  const allQueriesSucceeded = collectionQueries.every(result => result.isSuccess);
-
-  const collectionList = collectionQueries.map(({ data }) => ({
-    ...data,
-    modified_at: collection?.filter(item => item.id === data?.id)[0]?.modified_at,
-  }));
+  const collectionList = collectionQueries.map(
+    ({ data }) =>
+      data !== undefined && {
+        ...data,
+        modified_at: collection.filter(item => item.id === data.id)[0].modified_at,
+      }
+  );
 
   const selectItem = e => {
     setSelectedItem(e);
@@ -41,7 +42,9 @@ const Collection = ({ setImgSrc }) => {
       ) : (
         <>
           <Accordion variant="separated" w="100%" onChange={selectItem} value={selectedItem} mb={10}>
-            {allQueriesSucceeded && collectionList?.map(item => <Item key={item.id} item={item} />)}
+            {collectionList?.map(item => (
+              <Item key={item.id} item={item} />
+            ))}
           </Accordion>
           <PaginationGoup total={total} activePage={activePage} setActivePage={setActivePage} />
         </>
