--- src/components/myPage/SuggestedSubscription/index.js
@@ -1,2 +1,0 @@
-export { default as SuggestedSubscription } from './SuggestedSubscription';
-export { default as Content } from './Content';
